refactor: Consolidate token management into SessionRepository
- Removed TokenRepository and its corresponding interface
- Migrated token-related methods from TokenRepository to SessionRepository
- Updated AuthService and TokenService to use SessionRepository for token operations
- Added SuccessResponse contract class for common success responses
- Updated Services configuration to remove ITokenRepository registration
