+++ API/Contracts/Responses/Common/SuccessResponse.cs
@@ -1,0 +1,4 @@
+
+namespace API.Contracts.Responses.Common;
+
+public class SuccessResponse : ResponseBase { }
