+++ API/Repositories/SessionRepository.cs
@@ -33,7 +33,9 @@ public class SessionRepository : ISessionRepository
     {
         try
         {
-            return await _Context.Sessions.FindAsync(sessionId);
+            return await _Context.Sessions
+                .Include(s => s.Tokens)
+                .FirstOrDefaultAsync(s => s.Id == sessionId);
         }
         catch (Exception)
         {
@@ -99,6 +101,55 @@ public class SessionRepository : ISessionRepository
         catch (Exception)
         {
             throw;
+        }
+    }
+
+    public async Task AddToken(Token token)
+    {
+        try
+        {
+            await _Context.Tokens.AddAsync(token);
+            await _Context.SaveChangesAsync();
+        }
+        catch (Exception ex)
+        {
+            throw new Exception(ex.Message);
+        }
+    }
+
+    public async Task<Token?> GetTokenByTokenString(string tokenString)
+    {
+        try
+        {
+            return await _Context.Tokens.FirstOrDefaultAsync(t => t.TokenString == tokenString);
+        }
+        catch (Exception ex)
+        {
+            throw new Exception(ex.Message);
+        }
+    }
+
+    public async Task<Token?> GetTokenByUserId(long userId)
+    {
+        try
+        {
+            return await _Context.Tokens.FirstOrDefaultAsync(t => t.UserId == userId);
+        }
+        catch (Exception ex)
+        {
+            throw new Exception(ex.Message);
+        }
+    }
+
+    public async Task<IEnumerable<Token>> GetSessionTokensBySessionId(long sessionId)
+    {
+        try
+        {
+            return await _Context.Tokens.Where(t => t.SessionId == sessionId).ToListAsync();
+        }
+        catch (Exception ex)
+        {
+            throw new Exception(ex.Message);
         }
     }
 }
